Use a class for links
